fix: resolve core types layering and refresh dependency maps
