+++ .tickets/tun-3cde.md
@@ -1,0 +1,18 @@
+---
+id: tun-3cde
+status: open
+deps: [tun-48fe]
+links: []
+created: 2026-01-27T06:20:46Z
+type: task
+priority: 2
+tags: [docs, verification]
+---
+# Verify tests and update DEPENDENCY_MAP.md
+
+Run pytest, update docs/architecture/DEPENDENCY_MAP.md to reflect new structure
+
+## Acceptance Criteria
+
+Tests pass, documentation updated
+
