+++ .tickets/tun-3f92.md
@@ -1,0 +1,18 @@
+---
+id: tun-3f92
+status: closed
+deps: []
+links: []
+created: 2026-01-27T06:20:44Z
+type: task
+priority: 1
+tags: [architecture, types]
+---
+# Create core/types/ directory structure
+
+Create src/tunacode/core/types/ directory with __init__.py placeholder
+
+## Acceptance Criteria
+
+Directory exists with empty __init__.py
+
