+++ .tickets/tun-979a.md
@@ -1,0 +1,18 @@
+---
+id: tun-979a
+status: in_progress
+deps: [tun-3f92]
+links: []
+created: 2026-01-27T06:20:44Z
+type: task
+priority: 1
+tags: [architecture, types]
+---
+# Move 9 core-specific types to core/types/
+
+Move ResponseState, AgentState, StateManagerProtocol, SessionStateProtocol, ConversationState, TaskState, RuntimeState, UsageState, ToolCallRegistry from types/ to core/types/
+
+## Acceptance Criteria
+
+All 9 types importable from tunacode.core.types
+
