+++ src/tunacode/core/shared_types.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from tunacode.types import (
+from tunacode.types import (  # noqa: F401
     ModelName,
     ToolArgs,
     ToolCallback,
@@ -11,7 +11,7 @@ from tunacode.types import (
     ToolStartCallback,
     UserConfig,
 )
-from tunacode.types.canonical import UsageMetrics
+from tunacode.types.canonical import UsageMetrics  # noqa: F401
 
 __all__: list[str] = [
     "ModelName",
