+++ src/tunacode/core/types/tool_registry.py
