+++ src/tunacode/types/dataclasses.py
@@ -4,13 +4,6 @@ Contains structured data types used throughout the application.
 """
 
 from dataclasses import dataclass
-from enum import Enum
-from typing import TYPE_CHECKING
-
-from tunacode.types.callbacks import ProcessRequestCallback
-
-if TYPE_CHECKING:
-    from tunacode.types.state import StateManagerProtocol
 
 
 @dataclass
@@ -33,33 +26,6 @@ ModelRegistry = dict[str, ModelConfig]
 
 
 @dataclass
-class ResponseState:
-    """Track whether a user visible response was produced."""
-
-    has_user_response: bool = False
-    has_final_synthesis: bool = False
-    task_completed: bool = False
-    awaiting_user_guidance: bool = False
-
-
-class AgentState(Enum):
-    """Agent loop states for enhanced completion detection."""
-
-    USER_INPUT = "user_input"
-    ASSISTANT = "assistant"
-    TOOL_EXECUTION = "tool_execution"
-    RESPONSE = "response"
-
-
-@dataclass
-class CommandContext:
-    """Context passed to command handlers."""
-
-    state_manager: "StateManagerProtocol"
-    process_request: ProcessRequestCallback | None = None
-
-
-@dataclass
 class TokenUsage:
     """Token usage for a request."""
 
@@ -79,12 +45,9 @@ class CostBreakdown:
 
 
 __all__ = [
-    "AgentState",
-    "CommandContext",
     "CostBreakdown",
     "ModelConfig",
     "ModelPricing",
     "ModelRegistry",
-    "ResponseState",
     "TokenUsage",
 ]
